Moving ParameterSignature to a separate file
