+++ Sources/SwiftRewriterLib/Structures/FunctionSignature.swift
@@ -99,24 +99,6 @@ public struct FunctionSignature: Equatable, Codable {
     }
 }
 
-public struct ParameterSignature: Equatable, Codable {
-    public var label: String?
-    public var name: String
-    public var type: SwiftType
-    
-    public init(name: String, type: SwiftType) {
-        self.label = name
-        self.name = name
-        self.type = type
-    }
-    
-    public init(label: String?, name: String, type: SwiftType) {
-        self.label = label
-        self.name = name
-        self.type = type
-    }
-}
-
 public extension Sequence where Element == ParameterSignature {
     
     public func argumentLabels() -> [String?] {
